marServer/requirements.txt
django-cors-headers
django
django-rest-knox

psycopg2-binary
django-extensions

pandas
matplotlib
